Updated MGSKosmicTaskRunner framework to 32/64 bit and fixed warnings.

+++ KosmicTask/KosmicTask/source/mother task/MGSScriptRunner.m
@@ -13,7 +13,7 @@
 #import "MGSScriptRunnerApplication.h"
 
 
-#define OSStatusLog(status) NSLog(@"OSStatus error: %lx - reason: %s", status, GetMacOSStatusErrorString(status))
+#define OSStatusLog(status) NSLog(@"OSStatus error: %lx - reason: %s", (long)status, GetMacOSStatusErrorString(status))
 
 static BOOL isForegroundApplication = NO;
